+++ project_backend/Interfaces/IUserService.cs
@@ -2,7 +2,7 @@
 {
     public interface IUserService
     {
-        public int AddUserBalance(int userId, int balance);
+        public decimal AddUserBalance(int userId, decimal balance);
         public void PurchaseItem(int userId, string itemName, int quantityToBuy);
     }
 }
